pulls zk related functionality into separate command.
